* Average texture preparer added * Intensity texture preparers rearranged * float support for slice viewer added * OpenGL type query moved to separate source file (project file update needed for VS!) * Makefiles updated

+++ src/GUI/widgets/m4dGUISliceViewerWidget.h
@@ -778,7 +778,7 @@ protected:
     /**
      * The value of the recently picked voxel's color.
      */
-    int64					_colorPicked;
+    float64					_colorPicked;
 
     /**
      * The position on display of the recently picked voxel.
